feat: move etapa to core
