+++ central-oon-core-backend/src/controllers/etapa.js
@@ -1,27 +1,20 @@
-const EtapaService = require("../../services/etapa");
-const {
-  helpers: { sendResponse, sendPaginatedResponse },
-} = require("central-oon-core-backend");
+const EtapaService = require('../services/etapa');
+const { sendResponse, sendPaginatedResponse } = require('../utils/helpers');
 
 const criarEtapa = async (req, res) => {
   const etapa = await EtapaService.criar({ etapa: req.body });
-  sendResponse({
-    res,
-    statusCode: 201,
-    etapa,
-  });
+  sendResponse({ res, statusCode: 201, etapa });
 };
 
 const listarEtapas = async (req, res) => {
   const { pageIndex, pageSize, searchTerm, ...rest } = req.query;
 
-  const { etapas, limite, totalDeEtapas, page } =
-    await EtapaService.listarComPaginacao({
-      filtros: rest,
-      pageIndex,
-      pageSize,
-      searchTerm,
-    });
+  const { etapas, limite, totalDeEtapas, page } = await EtapaService.listarComPaginacao({
+    filtros: rest,
+    pageIndex,
+    pageSize,
+    searchTerm,
+  });
 
   sendPaginatedResponse({
     res,
@@ -40,33 +33,17 @@ const listarEtapasAtivasPorEsteira = async (req, res) => {
   const etapas = await EtapaService.listarEtapasAtivasPorEsteira({
     esteira: req.params.esteira,
   });
-  sendResponse({
-    res,
-    statusCode: 200,
-    etapas,
-  });
+  sendResponse({ res, statusCode: 200, etapas });
 };
 
 const atualizarEtapa = async (req, res) => {
-  const etapa = await EtapaService.atualizar({
-    etapa: req.body,
-    id: req.params.id,
-  });
-
-  sendResponse({
-    res,
-    statusCode: 200,
-    etapa,
-  });
+  const etapa = await EtapaService.atualizar({ etapa: req.body, id: req.params.id });
+  sendResponse({ res, statusCode: 200, etapa });
 };
 
 const excluir = async (req, res) => {
   const etapa = await EtapaService.excluir({ id: req.params.id });
-  sendResponse({
-    res,
-    statusCode: 200,
-    etapa,
-  });
+  sendResponse({ res, statusCode: 200, etapa });
 };
 
 module.exports = {
