+++ central-oon-core-backend/src/index.js
@@ -15,15 +15,19 @@ const Log = require('./models/Log');
 const ControleAlteracao = require('./models/ControleAlteracao');
 const Importacao = require('./models/Importacao');
 const Lista = require('./models/Lista');
+const Etapa = require('./models/Etapa');
 const controleAlteracaoController = require('./controllers/controleAlteracao');
 const importacaoController = require('./controllers/importacao');
 const listaController = require('./controllers/lista');
+const etapaController = require('./controllers/etapa');
 const controleAlteracaoRouter = require('./routers/controleAlteracaoRouter');
 const importacaoRouter = require('./routers/importacaoRouter');
 const listaRouter = require('./routers/listaRouter');
+const etapaRouter = require('./routers/etapaRouter');
 const ControleAlteracaoService = require('./services/controleAlteracao');
 const ImportacaoService = require('./services/importacao');
 const ListaService = require('./services/lista');
+const EtapaService = require('./services/etapa');
 const { registrarAcao } = require('./services/controleService');
 const { sendErrorResponse } = require('./utils/response');
 const helpers = require('./utils/helpers');
@@ -49,15 +53,19 @@ module.exports = {
     ControleAlteracao,
     Importacao,
     Lista,
+    Etapa,
     controleAlteracaoController,
     importacaoController,
     listaController,
+    etapaController,
     ControleAlteracaoService,
     ImportacaoService,
     ListaService,
+    EtapaService,
     controleAlteracaoRouter,
     importacaoRouter,
     listaRouter,
+    etapaRouter,
     sendErrorResponse,
     helpers,
     excel,
