+++ src/routers/seedRouter.js
@@ -5,7 +5,7 @@ const BaseOmie = require("../models/BaseOmie");
 const Usuario = require("../models/Usuario");
 // const Banco = require("../models/Banco");
 // const Estado = require("../models/Estado");
-const Etapa = require("../models/Etapa");
+const { Etapa } = require("central-oon-core-backend");
 const Sistema = require("../models/Sistema");
 const ListaOmie = require("../models/ListaOmie");
 const Assistente = require("../models/Assistente");
