+++ src/services/documentoFiscal/index.js
@@ -7,7 +7,7 @@ const Arquivo = require("../../models/Arquivo");
 const { criarNomePersonalizado } = require("../../utils/formatters");
 const ArquivoNaoEncontradoError = require("../errors/arquivo/arquivoNaoEncontradoError");
 const DocumentoFiscalNaoEncontradaError = require("../errors/documentoFiscal/documentoFiscalNaoEncontradaError");
-const EtapaService = require("../etapa");
+const { EtapaService } = require("central-oon-core-backend");
 const ServicoTomadoTicket = require("../../models/ServicoTomadoTicket");
 const Servico = require("../../models/Servico");
 const { default: mongoose } = require("mongoose");
